# api/docs/swagger.yaml
basePath: /v1
definitions:
  category.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        description: UUID
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  category.CreateCategoryRequest:
    properties:
      name:
        type: string
    type: object
  category.DeleteCategoryResponse:
    properties:
      message:
        type: string
    type: object
  category.GetAllCategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/category.Category'
        type: array
    type: object
  category.UpdateCategoryRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  comment.Comment:
    properties:
      body:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        description: UUID
        type: string
      post_id:
        description: UUID
        type: string
      updated_at:
        type: string
      user_id:
        description: UUID
        type: string
    type: object
  comment.CreateCommentRequest:
    properties:
      body:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
  comment.DeleteCommentResponse:
    properties:
      message:
        type: string
    type: object
  comment.GetAllCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
    type: object
  comment.UpdateCommentRequest:
    properties:
      body:
        type: string
      id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a forum API gateway service.
  title: Forum API Gateway
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all categories with optional pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of categories per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.GetAllCategoriesResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - category
    post:
      consumes:
      - application/json
      description: Create a new category with given information
      parameters:
      - description: Category information
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - category
    put:
      consumes:
      - application/json
      description: Update a category with the given information
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category information
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Category item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a category by its ID
      tags:
      - category
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by its unique ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.DeleteCategoryResponse'
        "404":
          description: Category item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a category by its ID
      tags:
      - category
    get:
      consumes:
      - application/json
      description: Retrieve a category by its unique ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Category item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a category by its ID
      tags:
      - category
  /comments:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all comments with optional filtering and pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of comments per page
        in: query
        name: limit
        type: integer
      - description: Filter by post ID
        in: query
        name: post_id
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.GetAllCommentsResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create a new comment with given information
      parameters:
      - description: Comment information
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/comment.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new comment
      tags:
      - comment
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by its unique ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.DeleteCommentResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Comment item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a comment by its ID
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Retrieve a comment by its unique ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Comment item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a comment by its ID
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Update a comment with the given information
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment information
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/comment.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Comment item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a comment by its ID
      tags:
      - comment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"